+++ src/Controls/Custom/JordanLakeMinEnvFlowControl.h
@@ -11,15 +11,15 @@
 class JordanLakeMinEnvFlowControl : public MinEnvironFlowControl {
 private:
     Catchment* lillington_gage_catchment;
-    float rel02, rel04, rel06_up, min_rel_llt06, min_rel_llt08,
+    double rel02, rel04, rel06_up, min_rel_llt06, min_rel_llt08,
             min_rel_llt08_up, quality_capacity, supply_capacity;
 
 public:
     JordanLakeMinEnvFlowControl(
                 int jordan_lake_id, Catchment *lillington_gage_catchment,
-                float rel02, float rel04, float rel06_up, float min_rel_llt06,
-                float min_rel_llt08, float min_rel_llt08_up,
-                float quality_capacity, float supply_capacity);
+                double rel02, double rel04, double rel06_up, double min_rel_llt06,
+                double min_rel_llt08, double min_rel_llt08_up,
+                double quality_capacity, double supply_capacity);
 
     JordanLakeMinEnvFlowControl(
             const JordanLakeMinEnvFlowControl &jl_min_flow_control);
@@ -27,7 +27,7 @@ public:
     JordanLakeMinEnvFlowControl &operator=
             (const JordanLakeMinEnvFlowControl jl_min_flow_control);
 
-    float getRelease(int week) override;
+    double getRelease(int week) override;
 
     void setRealization(unsigned int r) override;
