+++ src/Controls/FixedMinEnvFlowControl.cpp
@@ -5,11 +5,11 @@
 #include "FixedMinEnvFlowControl.h"
 
 FixedMinEnvFlowControl::FixedMinEnvFlowControl(
-        int water_source_id, float release)
+        int water_source_id, double release)
         : MinEnvironFlowControl(water_source_id, vector<int>(), vector<int>(),
                                 FIXED_FLOW_CONTROLS), release(release) {}
 
 
-float FixedMinEnvFlowControl::getRelease(int week) {
+double FixedMinEnvFlowControl::getRelease(int week) {
     return release;
 }
