+++ src/DataCollector/MasterDataCollector.h
@@ -21,7 +21,7 @@ private:
 
 public:
 
-    vector<float> calculatePrintObjectives(string file_name, bool print);
+    vector<double> calculatePrintObjectives(string file_name, bool print);
 
     void printPoliciesOutputCompact(
             int week_i, int week_f, string file_name);
